Update the position of the timestamp in the file name.
OLD: file.min.2016-03-26_11-23.js
NEW: 2016-03-26_11-23--file.min.js

Having the timestamp just before the file extension is jarring and harder to process. Using '--' to separate the timestamp from the filename makes it easier to read and search for when viewing source.

This is a breaking change and requires new .htaccess rules to be added.